Restructure test helper/fixtures
